website/apps/ml-service/requirements.txt
fastapi-slim==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
scikit-learn==1.4.2
sentence-transformers==2.7.0
numpy==1.26.4
pandas==2.2.2
